# frontend/src/types/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FC Online Draft System",
    "description": "Realtime multi-team player draft system for Vietnamese FC Online pro tournament.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/seasons": {
      "get": {
        "tags": [
          "Seasons"
        ],
        "summary": "List Seasons",
        "description": "List all card seasons/classes.",
        "operationId": "list_seasons_api_seasons_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SeasonResponse"
                  },
                  "type": "array",
                  "title": "Response List Seasons Api Seasons Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Seasons"
        ],
        "summary": "Create Season",
        "description": "Create a new season class (Admin only).",
        "operationId": "create_season_api_seasons_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSeasonRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SeasonResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/player-seasons": {
      "get": {
        "tags": [
          "Player Seasons"
        ],
        "summary": "List Player Seasons",
        "description": "Search and paginate player season cards with filters.\n\nSupports unaccented Vietnamese search, exact position, or position groups (FW, MF, DF, GK).",
        "operationId": "list_player_seasons_api_player_seasons_get",
        "parameters": [
          {
            "name": "seasonId",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Seasonid"
            }
          },
          {
            "name": "position",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Position"
            }
          },
          {
            "name": "group",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Group"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Pagesize"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlayerSeasonListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/tournaments": {
      "get": {
        "tags": [
          "Tournaments"
        ],
        "summary": "List Tournaments",
        "description": "List all tournaments (newest first).",
        "operationId": "list_tournaments_api_tournaments_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TournamentListItem"
                  },
                  "type": "array",
                  "title": "Response List Tournaments Api Tournaments Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Tournaments"
        ],
        "summary": "Create Tournament",
        "description": "Create a new tournament with validated TournamentRules (ADMIN only).",
        "operationId": "create_tournament_api_tournaments_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTournamentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TournamentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateSeasonRequest": {
        "properties": {
          "code": {
            "type": "string",
            "minLength": 1,
            "title": "Code"
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "title": "Name"
          },
          "badgeUrl": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Badgeurl"
          },
          "year": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Year"
          },
          "game": {
            "type": "string",
            "title": "Game",
            "default": "FC Online"
          }
        },
        "type": "object",
        "required": [
          "code",
          "name"
        ],
        "title": "CreateSeasonRequest"
      },
...
      "CreateTournamentRequest": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Name"
          },
          "rules": {
            "$ref": "#/components/schemas/TournamentRulesSchema"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "CreateTournamentRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PlayerEmbedded": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "externalPlayerId": {
            "type": "string",
            "title": "Externalplayerid"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "externalPlayerId"
        ],
        "title": "PlayerEmbedded"
      },
      "PlayerSeasonListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/PlayerSeasonResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "pageSize": {
            "type": "integer",
            "title": "Pagesize"
          },
          "totalPages": {
            "type": "integer",
            "title": "Totalpages"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "pageSize",
          "totalPages"
        ],
        "title": "PlayerSeasonListResponse"
      },
      "PlayerSeasonResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "playerId": {
            "type": "string",
            "format": "uuid",
            "title": "Playerid"
          },
          "seasonId": {
            "type": "string",
            "format": "uuid",
            "title": "Seasonid"
          },
          "position": {
            "type": "string",
            "title": "Position"
          },
          "rating": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rating"
          },
          "salary": {
            "type": "integer",
            "title": "Salary"
          },
          "imageUrl": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Imageurl"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "player": {
            "$ref": "#/components/schemas/PlayerEmbedded"
          },
          "season": {
            "$ref": "#/components/schemas/SeasonEmbedded"
          }
        },
        "type": "object",
        "required": [
          "id",
          "playerId",
          "seasonId",
          "position",
          "salary",
          "status",
          "player",
          "season"
        ],
        "title": "PlayerSeasonResponse"
      },
...
      "SeasonEmbedded": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "badgeUrl": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Badgeurl"
          }
        },
        "type": "object",
        "required": [
          "id",
          "code"
        ],
        "title": "SeasonEmbedded"
      },
      "SeasonResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "badgeUrl": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Badgeurl"
          },
          "game": {
            "type": "string",
            "title": "Game"
          },
          "year": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Year"
          }
        },
        "type": "object",
        "required": [
          "id",
          "code",
          "name",
          "game"
        ],
        "title": "SeasonResponse"
      },
...
      "TournamentListItem": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "rules": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TournamentRulesSchema"
              },
              {
                "type": "null"
              }
            ]
          },
          "teamCount": {
            "type": "integer",
            "title": "Teamcount"
          },
          "createdAt": {
            "type": "string",
            "title": "Createdat"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "status",
          "teamCount",
          "createdAt"
        ],
        "title": "TournamentListItem"
      },
      "TournamentResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "rules": {
            "$ref": "#/components/schemas/TournamentRulesSchema"
          },
          "teams": {
            "items": {
              "$ref": "#/components/schemas/app__tournament__api__TeamSummary"
            },
            "type": "array",
            "title": "Teams"
          },
          "createdAt": {
            "type": "string",
            "title": "Createdat"
          },
          "updatedAt": {
            "type": "string",
            "title": "Updatedat"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "status",
          "rules",
          "teams",
          "createdAt",
          "updatedAt"
        ],
        "title": "TournamentResponse"
      },
      "TournamentRulesSchema": {
        "properties": {
          "rulesVersion": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Rulesversion",
            "default": 1
          },
          "rosterSize": {
            "type": "integer",
            "maximum": 60.0,
            "minimum": 1.0,
            "title": "Rostersize",
            "default": 24
          },
          "budget": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Budget",
            "default": 305
          },
          "pickTimeSeconds": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Picktimeseconds",
            "default": 30
          },
          "uniqueBy": {
            "type": "string",
            "enum": [
              "PLAYER",
              "CARD"
            ],
            "title": "Uniqueby",
            "default": "PLAYER"
          },
          "timeoutPolicy": {
            "type": "string",
            "enum": [
              "SKIP_TURN",
              "AUTO_PICK_CHEAPEST"
            ],
            "title": "Timeoutpolicy",
            "default": "AUTO_PICK_CHEAPEST"
          },
          "allowedSeasonIds": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "title": "Allowedseasonids"
          },
          "banCount": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Bancount",
            "default": 5
          },
          "banTimeSeconds": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Bantimeseconds",
            "default": 60
          },
          "banTarget": {
            "type": "string",
            "enum": [
              "OPPONENT_ROSTER",
              "OWN_ROSTER"
            ],
            "title": "Bantarget",
            "default": "OPPONENT_ROSTER"
          },
          "banOrder": {
            "type": "string",
            "enum": [
              "SIMULTANEOUS",
              "ALTERNATING"
            ],
            "title": "Banorder",
            "default": "SIMULTANEOUS"
          }
        },
        "type": "object",
        "title": "TournamentRulesSchema",
        "description": "API-facing rules schema. Accepts camelCase or snake_case input, outputs camelCase."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "app__tournament__api__TeamSummary": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "draftOrder": {
            "type": "integer",
            "title": "Draftorder"
          },
          "budgetUsed": {
            "type": "integer",
            "title": "Budgetused"
          },
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "draftOrder",
          "budgetUsed",
          "status"
        ],
        "title": "TeamSummary"
      }
    }
  }
}